e2e: add basic testing (#15)
This adds some very basic e2e tests. I've picked babashka because
I started with Bash and then couldn't take it anymore. Let's see how
it works out for us here.

For some reason, this didn't run on the PR, so I've merged it as-is, and
will try to get it to run (and fix whatever is broken) in a subsequent PR.

Signed-off-by: Stephan Renatus <stephan@styra.com>